WebpageDev/requirements.txt

Flask==2.2.3
Flask-SocketIO==5.3.3
paho-mqtt==1.6.1

